requirements.txt
django>=5.2,<5.3
google-generativeai
numpy
Pillow
